+++ back/src/bsda/converter.ts
@@ -243,8 +243,29 @@ export function expandBsdaFromElastic(bsda: BsdaForElastic): GraphqlBsda {
   return {
     ...expanded,
     groupedIn,
-    forwardedIn
+    forwardedIn,
+    metadata: {
+      latestRevision: computeLatestRevision(bsda.bsdaRevisionRequests) as any
+    }
   };
+}
+
+export function computeLatestRevision(
+  revisionRequests: BsdaRevisionRequest[] | null
+) {
+  if (!revisionRequests || revisionRequests.length === 0) {
+    return null;
+  }
+
+  return revisionRequests.reduce((latestRevision, currentRevision) => {
+    if (
+      !latestRevision ||
+      currentRevision.updatedAt > latestRevision.updatedAt
+    ) {
+      return currentRevision;
+    }
+    return latestRevision;
+  });
 }
 
 export function flattenBsdaInput(formInput: BsdaInput) {