+++ front/src/Apps/common/Components/layout/Header.tsx
@@ -4,8 +4,7 @@ import {
   Link,
   matchPath,
   useLocation,
-  generatePath,
-  useMatch
+  generatePath
 } from "react-router-dom";
 
 import { localAuthService } from "../../../../login/auth.service";
@@ -24,8 +23,7 @@ import {
 } from "../../../../common/config";
 import styles from "./Header.module.scss";
 import { useMedia } from "../../../../common/use-media";
-import { DashboardTabs } from "../../../../dashboard/DashboardTabs";
-import { default as DashboardTabsV2 } from "../../../Dashboard/Components/DashboardTabs/DashboardTabs";
+import { default as DashboardTabs } from "../../../Dashboard/Components/DashboardTabs/DashboardTabs";
 
 export const GET_ME = gql`
   {
@@ -35,6 +33,7 @@ export const GET_ME = gql`
         id
         name
         givenName
+        userRole
         orgId
         companyTypes
         userPermissions
@@ -51,7 +50,6 @@ export const GET_ME = gql`
 function MobileSubNav({ currentSiret }) {
   const { updatePermissions } = usePermissions();
   const { error, data } = useQuery<Pick<Query, "me">>(GET_ME, {});
-  const isV2Routes = !!useMatch("/v2/dashboard/*");
 
   useEffect(() => {
     if (data) {
@@ -60,7 +58,10 @@ function MobileSubNav({ currentSiret }) {
         company => company.orgId === currentSiret
       );
       if (currentCompany) {
-        updatePermissions(currentCompany.userPermissions);
+        updatePermissions(
+          currentCompany.userPermissions,
+          currentCompany.userRole!
+        );
       }
     }
   }, [updatePermissions, data, currentSiret]);
@@ -79,10 +80,8 @@ function MobileSubNav({ currentSiret }) {
     return null;
   }
 
-  return !isV2Routes ? (
+  return (
     <DashboardTabs currentCompany={currentCompany} companies={companies} />
-  ) : (
-    <DashboardTabsV2 currentCompany={currentCompany} companies={companies} />
   );
 }
 
@@ -118,19 +117,9 @@ const getMenuEntries = (isAuthenticated, isAdmin, currentSiret) => {
 
   const connected = [
     {
-      caption: "Mon espace",
+      caption: "Mes bordereaux",
       href: currentSiret
         ? generatePath(routes.dashboard.index, {
-            siret: currentSiret
-          })
-        : "/",
-
-      navlink: true
-    },
-    {
-      caption: "Mes bordereaux 🆕",
-      href: currentSiret
-        ? generatePath(routes.dashboardv2.index, {
             siret: currentSiret
           })
         : "/",
@@ -241,26 +230,10 @@ export default function Header({
     location.pathname
   );
 
-  const matchDashboardV2 = matchPath(
-    {
-      path: routes.dashboardv2.index,
-      caseSensitive: false,
-      end: false
-    },
-    location.pathname
-  );
-
   const menuClass = menuHidden && isMobile ? styles.headerNavHidden : "";
 
   // Catching siret from url when not available from props (just after login)
-  let currentSiret = matchDashboard?.params["siret"] || defaultOrgId;
-
-  if (matchDashboardV2) {
-    currentSiret =
-      matchDashboard?.params["siret"] ||
-      matchDashboardV2?.params["siret"] ||
-      defaultOrgId;
-  }
+  const currentSiret = matchDashboard?.params["siret"] || defaultOrgId;
 
   const menuEntries = getMenuEntries(isAuthenticated, isAdmin, currentSiret);