+++ front/src/form/common/components/company/ExtraEuropeanCompanyManualInput.tsx
@@ -1,8 +1,10 @@
-import RedErrorMessage from "../../../../common/components/RedErrorMessage";
-import { Field, useField } from "formik";
+import { Field, useField, useFormikContext } from "formik";
 import React from "react";
 
 import { FormCompany } from "@td/codegen-ui";
+import CountrySelector from "./CountrySelector";
+import RedErrorMessage from "../../../../common/components/RedErrorMessage";
+import Tooltip from "../../../../common/components/Tooltip";
 
 interface ExtraEuropeanCompanyManualProps {
   name: string;
@@ -18,17 +20,29 @@ export default function ExtraEuropeanCompanyManualInput({
   onExtraEuropeanCompanyId
 }: ExtraEuropeanCompanyManualProps) {
   const [field] = useField<FormCompany>({ name });
+  const { setFieldValue } = useFormikContext();
 
   return (
-    <div className="tw-my-6">
+    <div>
       <h4 className="form__section-heading">Entreprise extra-européenne</h4>
       <div className="form__row">
         <label>
           Identifiant de l'entreprise
+          <Tooltip msg="À renseigner si numéro de TVA inexistant" />
           <input
             type="text"
             value={extraEuropeanCompanyId!}
             onChange={e => onExtraEuropeanCompanyId(e.target.value)}
+            className="td-input"
+          />
+        </label>
+        <label>
+          Numéro de TVA (optionnel)
+          <Tooltip msg="À renseigner si le numéro de TVA n'est pas reconnu dans le champ de recherche" />
+          <input
+            type="text"
+            value={field.value.vatNumber!}
+            name={`${field.name}.vatNumber`}
             className="td-input"
           />
         </label>
@@ -58,12 +72,9 @@ export default function ExtraEuropeanCompanyManualInput({
         <RedErrorMessage name={`${field.name}.address`} />
         <label>
           Pays de l'entreprise
-          <Field
-            type="text"
-            className="td-input"
-            name={`${field.name}.country`}
-            optional={optional}
-            placeholder="code pays deux lettres majuscules norme ISO 3166-1"
+          <CountrySelector
+            value={field.value.country!}
+            onChange={value => setFieldValue(`${field.name}.country`, value)}
           />
         </label>
         <RedErrorMessage name={`${field.name}.country`} />